Finished collision code for the ball and the blocks and added block sprites
Currently the ball calls the Destroy() function to get rid of block but
I will change this to call a Damage() function so that I have some
flexibility and can work with different block types. Otherwise the core
functionality of the game is finished, leaving me to start work on a
level framework.